feat(rfc-0207): A1 — fix CAG classification (Set.has exact -> substring)
BUG #1 fix: flip climatizacao identifier matching from EXACT (identifierEquals,
Set.has) to SUBSTRING (identifierContains, .includes) in the DEFAULT seed, both
the conditional (BOMBA/MOTOR) and identifier-fallback branches. Unifies with
equipmentCategory.js (already .includes()). The resolver already supported
identifierContains (built in A0), so the fix is data-only.

Deliberate behavior change — replaces the A0 equivalence golden with a migration
snapshot. resolveGroup (columns) stays equivalent to legacy; resolveCategory
moves EXACTLY the reviewed devices, nothing else:

  device identifier | deviceProfile | before    -> after
  CAG 01            | BOMBA         | outros    -> climatizacao
  BOMBA CAG 2       | BOMBA         | outros    -> climatizacao
  CAG PRINCIPAL     | MOTOR         | outros    -> climatizacao
  CAG 01            | (none)        | outros    -> climatizacao   (identifier fallback)

Migration test 6/6 green (asserts no surprise moves). build:tsup ESM+DTS pass.

Deferred to follow-up sub-PRs: buildSummary subcategorization (bug #2),
inferLabelWidget (:866) substring path, equipmentCategory.js parity decision.
